Sound Effects
On Tino3 ,I added the some of the sound effects, however I was receiving some errors on SimpleMovement.cs on lines 97 to 101, so I commented them out to test out my code. But I left it back to how it originially was set up. However, comment them out to test the sound if you guys would like.

There are still a few of sound effects left over to edit